Create initial picker image component
Create initial picker image component in the create event component
