+++ src/app/events/event-create/image-picker/image-picker.component.html
@@ -1,5 +1,5 @@
 <div class="image-picker-wrapper">
-  <div class="image-picker" (click)="onImageSelect()">
+  <div class="image-picker" (click)="onAttachPictureClick()">
     <img
       src="../../../assets/icon/camera.png"
       alt="image-picker-camera"
