+++ src/app/events/event-create/image-picker/image-picker.component.scss
