+++ src/app/events/event-create/image-picker/image-picker.component.spec.ts
