+++ src/app/events/event-create/image-picker/image-picker.component.ts
@@ -1,0 +1,41 @@
+import { Component, EventEmitter, OnInit, Output } from '@angular/core';
+import {
+  Camera,
+  CameraResultType,
+  CameraSource,
+  Photo,
+} from '@capacitor/camera';
+
+@Component({
+  selector: 'app-image-picker',
+  templateUrl: './image-picker.component.html',
+  styleUrls: ['./image-picker.component.scss'],
+})
+export class ImagePickerComponent implements OnInit {
+  @Output() imagePicked = new EventEmitter<string>();
+  selectedImage: string = '';
+
+  constructor() {}
+
+  ngOnInit() {}
+
+  async onAttachPictureClick() {
+    console.log('dada');
+
+    const image = await Camera.getPhoto({
+      quality: 100,
+      allowEditing: false,
+      resultType: CameraResultType.Uri,
+      source: CameraSource.Photos,
+    });
+
+    if (image) {
+      this.imagePicked.emit(image.webPath!);
+      this.savePicture(image);
+    }
+  }
+
+  async savePicture(photo: Photo) {
+    console.log(photo);
+  }
+}
