--- src/shared/components/image-picker/image-picker.component.ts
@@ -1,16 +1,0 @@
-import { Component, OnInit } from '@angular/core';
-
-@Component({
-  selector: 'app-image-picker',
-  templateUrl: './image-picker.component.html',
-  styleUrls: ['./image-picker.component.scss'],
-})
-export class ImagePickerComponent implements OnInit {
-  constructor() {}
-
-  ngOnInit() {}
-
-  onImageSelect() {
-    console.log('dada');
-  }
-}
